Added a button that opens a MapView
I am using a test api key.
MapView opens in a new window and the size cannot be changed.
A static mapview ( when pressed will open a MapVIew) can be added as a container, working on adding that functionality in my next commit. The Frontend team can shape it how they would like.