Added OAuth profile creation and saving profile id to app(#33)
* Added disclaimer screen to the main flow

* Added split screen to Ride Screen

* Fixed the bug where clicking the submit button without actually changing
anything would erase the address instead of not changing anything.

* Fixed the setState bug!

* Added the disclaimer screen with accept button
1st time the app is opened it takes the user to
a disclaimer screen. After that first time, the
user is taken directly to the main screen. Will
change to OAuth later.

* Changed the state for the messages

* Added the auth screen to the first time login

* changed update url for new profile

* Added the edit schedule screen - currently working on it
Changed the disclaimer to go to main.dart after acceptance for easier testing
Added a initial profile screen file which is going to include the
editing buttons for the schedule and the address

* adde profile post method

* Added schedulemap variable for holding the final schedule
Added drop down menus for each day
Added a table format to the screen

* changed to id auto increments

* changed to id auto increments

* Removed the disclaimer

* Setting up initial authentication of user

* Merged the files from master
Added the edit schedule screen where the user can input their
schedule and it saves it in military time in a scheduleMap variable
with each day mapping to its schedule.
Also added getters and setters for every day's schedule

* Fixed the app drawer

* attempt to post monday, not working

* change to model not